Reverse Selection, Mark Selection Start/End, Pointer/PCM views, Decimal Ruler, Mini-Map drag scroll,

+++ Tools/ExpressionEvaluator/Src/EEExpEvalContainer.cpp
@@ -645,7 +645,7 @@ namespace ee {
 
 						ee::CVector * pThis = static_cast<ee::CVector *>(_rRight.u.poObj);
 						try {
-							psObj->Resize( pThis->Size().u.ui64Val );
+							psObj->Resize( size_t( pThis->Size().u.ui64Val ) );
 						}
 						catch ( ... ) {
 							DeallocateObject( psObj );
